Removed placeholder data
