+++ src/components/App.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+import "./App.css";
+
+function App() {
+  return <div className="App"></div>;
+}
+
+export default App;
